+++ packages/docs/src/content/examples/header-bar/usage.vue
@@ -36,6 +36,9 @@
 		];
 
 		options = {
+			booleans: [
+				'mobileVersion'
+			],
 			selects: {
 				theme: [
 					'default',